Move views mixins to internal
